LifeCycleTest : LifeCycleTest package packed.
- Packed following files into LifeCycleTest package
	- LifeCycleTest_eat.java
	- LifeCycleTest_getLifeCycleStatus.java
	- LifeCycleTest_haveSex.java
	- LifeCycleTest_isAdult.java
	- LifeCycleTest_isDead.java
	- LifeCycleTest_isHungry.java
	- LifeCycleTest_isOld.java
	- LifeCycleTest_isPregnant.java
	- LifeCycleTest_isYoung.java
	- LifeCycleTest_isYoungAndFat.java
	- LifeCycleTest_move.java
