+++ src/test/java/org/woehlke/computer/kurzweil/simulated/evolution/model/cell/LifeCycleTest/LifeCycleTest_haveSex.java
@@ -1,9 +1,10 @@
-package org.woehlke.computer.kurzweil.simulated.evolution.model.cell;
+package org.woehlke.computer.kurzweil.simulated.evolution.model.cell.LifeCycleTest;
 
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.jupiter.api.Assertions;
+import org.woehlke.computer.kurzweil.simulated.evolution.model.cell.LifeCycle;
 
 import static org.junit.Assert.*;
