+++ src/test/java/org/woehlke/computer/kurzweil/simulated/evolution/model/cell/LifeCycleTest/LifeCycleTest_isAdult.java
@@ -1,8 +1,9 @@
-package org.woehlke.computer.kurzweil.simulated.evolution.model.cell;
+package org.woehlke.computer.kurzweil.simulated.evolution.model.cell.LifeCycleTest;
 
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.jupiter.api.Assertions;
+import org.woehlke.computer.kurzweil.simulated.evolution.model.cell.LifeCycle;
 
 import static org.junit.Assert.*;
