+++ src/test/java/org/woehlke/computer/kurzweil/simulated/evolution/model/cell/LifeCycleTest/LifeCycleTest_isYoungAndFat.java
@@ -1,10 +1,15 @@
-package org.woehlke.computer.kurzweil.simulated.evolution.model.cell;
+package org.woehlke.computer.kurzweil.simulated.evolution.model.cell.LifeCycleTest;
 
+import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.jupiter.api.Assertions;
+import org.woehlke.computer.kurzweil.simulated.evolution.model.cell.LifeCycle;
 
-public class LifeCycleTest_isPregnant {
+import static org.junit.Assert.*;
+
+public class LifeCycleTest_isYoungAndFat {
+
     private LifeCycle lc1;
     private LifeCycle lc2;
     private LifeCycle lc3;
@@ -29,10 +34,11 @@ public class LifeCycleTest_isPregnant {
         lc4.setFat(lc4.getFAT_MINIMUM_FOR_SEX() + 1);
     }
     @Test
-    public void isPregnant() {
-        Assertions.assertTrue(!lc1.isPregnant());
-        Assertions.assertTrue(!lc2.isPregnant());
-        Assertions.assertTrue(!lc3.isPregnant());
-        Assertions.assertTrue(lc4.isPregnant());
+    public void isYoungAndFat() {
+        Assertions.assertTrue(!lc1.isYoungAndFat());
+        Assertions.assertTrue(lc2.isYoungAndFat());
+        Assertions.assertTrue(!lc3.isYoungAndFat());
+        Assertions.assertTrue(!lc4.isYoungAndFat());
     }
+
 }
